requirements.txt

python-telegram-bot==20.4
requests
